Implement KIF system actor steps supporting isolated controller test use cases
This pull request introduces a handful of methods onto `KIFSystemTestActor` that support using KIF to develop and test UIViewControllers in isolation.
The methods introduced provide a simple interface for instantiating a view controller, optionally yielding it to the test for configuration, and then
presenting it for subsequent interaction by the test case.

In general, this approach leverages KIF more as a functional testing harness than its traditional role in integration/acceptance tests. More specifically,
this approach is useful in a number of scenarios:

1. Doing view controller TDD as a separate concern from the rest of the app.
2. Developing view controllers that cannot yet be integrated into the main UI (perhaps due to missing design or external dependencies)
3. Testing view controllers in scenarios that are otherwise hard to recreate with the UI alone